ECO/paddle1.8/requirements.txt
Python>=3.7.0
PaddlePaddle>=1.8.0